requirements.txt
alembic==1.12.0
pytest==7.4.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
pytest-benchmark==4.0.0
pytest-postgresql==5.0.0
python-dotenv==1.0.0
PyYAML==6.0.1 